* string.c (rb_str_inspect): suppress a warning.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@26396 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ string.c
@@ -4101,7 +4101,7 @@ rb_str_inspect(VALUE str)
 	    if (p > prev) str_buf_cat(result, prev, p - prev);
             n = rb_enc_mbminlen(enc);
             if (pend < p + n)
-                n = pend - p;
+                n = (int)(pend - p);
             while (n--) {
                 snprintf(buf, CHAR_ESC_LEN, "\\x%02X", *p & 0377);
                 str_buf_cat(result, buf, strlen(buf));
